requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pandas
google-genai